numerical_summarizer_agent/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
curl_cffi==0.11.3
docstring_parser==0.16
Flask==3.1.1
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-aiplatform==1.88.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-pubsub==2.30.0
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.19.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
google-generativeai>=0.8.0
grpc-google-iam-v1==0.14.2
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
MarkupSafe==3.0.2
multitasking==0.0.11
narwhals==1.42.0
numpy==2.2.4
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==4.21.6
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yfinance==0.2.62
zipp==3.23.0
zstandard==0.23.0
